Add subreddit post count
