+++ devvit.yaml
@@ -1,2 +1,2 @@
 name: modmail-userinfo
-version: 1.5.6
+version: 1.5.6.1
